pyproject.toml
[tool.poetry]
name = "deepfoot"
version = "0.1.0"
description = ""
authors = ["Hatem Elseidy <hatem.faheem@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
opencv-python = "^4.10.0.84"
openai = "^1.43.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.6.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
